# Higher precision for floats in range [0,1)