First a disclaimer :).
With all the discussions happening around rand, I am not sure where to open a pull request.
Also it is a bit difficult for me to explain what I am doing here...

This commit changes the way floats are created with the distribution `Uniform01`.
This is not a method I have found anywhere else, but loosely inspired by
https://readings.owlfolio.org/2007/generating-pseudorandom-floating-point-values/ .

With this method floats have a much better precision near zero (9 extra bits of precision for f32, and 12 extra bits for f64).

It seems to have almost no influence on the benchmarks:
before:
`test distr_baseline          ... bench:       1,616 ns/iter (+/- 68) = 4950 MB/s`
after:
`test distr_baseline          ... bench:       1,621 ns/iter (+/- 30) = 4935 MB/s`