@dhardy 2017-08-23
Nice. I haven't had the time today but I'll take a look and merge.

I haven't focussed on changes like this since getting the main architecture right seemed like a priority, but this is definitely worth bringing up and mentioning on the main RFC thread.