@pitdicker 2017-08-23
Thank you, and there is no hurry for me.

I am now looking what the consequences are for other parts of rand that use this.

`range` for example looses most of the precision again if one of it's numbers is negative. I have fixed this, but not tested en benchmarked yet.
`open01` seems to rely on some precision also...

Tomorrow I will try to write something on the RFC tread.