@pitdicker 2017-08-25
A quick update:
`open01`, `closed01` and `range` now preserve the extra bits of precision near 0.0.

The benchmark `distr_range2_float` has regressed 40%. Moving the code conversion from uint to float out from `closed01`, `uniform01` and `uniform11` fixes this. Probably some problem with inlining?

This is a good chance to simplify the code to use constants, and wrap it in a macro so there is no duplication between f32 and f64.

Which is a long way of saying, please wait a few more days before reviewing :)