@pitdicker 2017-08-28
The code in `ziggurat` in `distributions` uses it's own method of converting a random int to a float.
It only uses as many bits of precision as bit in the floats fraction, and uses the rest to pick a row from the ziggurat. It is perfect to use the previous code from `Uniform01` here. The `distribution` benchmarks improve by 5~7%.

This got me thinking that it may be best to separate out the int to float conversions to an other file. It does not really fit in `uniform.rs`.

I'll work on it some more...