@dhardy 2017-08-28
Sure, putting the float conversions in a separate file makes sense (`uniform_float.rs` or `float.rs` or some such).

I'm not so sure about making `open01` etc. member functions on `u32` / `u64`. It may be useful to make these public, sure, but in general it seems a weird operation to make a member function.