@pitdicker 2017-08-28
You are right that making them member functions is not ideal. But I don't know enough rust yet for something better. Do you have a suggestion, or shall I just try a bit more?