@dhardy 2017-08-28
Simple functions (generic or with separate names for each type) would work. But I don't know if this would be better. As noted in my edit above, using member functions is probably fine actually.