@pitdicker 2017-08-28
Moving the functions to a separate file took care of them being public, so I was lucky :-).

I have tried to clean up de commit history.

Some benchmarks:
Before:
```
test distr_baseline          ... bench:       1,620 ns/iter (+/- 17) = 4938 MB/s
test distr_exp               ... bench:       6,286 ns/iter (+/- 131) = 1272 MB/s
test distr_gamma_large_shape ... bench:      18,190 ns/iter (+/- 317) = 439 MB/s
test distr_gamma_small_shape ... bench:      21,233 ns/iter (+/- 427) = 376 MB/s
test distr_log_normal        ... bench:       6,650 ns/iter (+/- 94) = 1203 MB/s
test distr_normal            ... bench:       6,651 ns/iter (+/- 99) = 1202 MB/s
test distr_range2_float      ... bench:       1,620 ns/iter (+/- 17) = 4938 MB/s
test distr_range2_int        ... bench:       3,102 ns/iter (+/- 16) = 2578 MB/s
test distr_range_float       ... bench:       1,620 ns/iter (+/- 17) = 4938 MB/s
test distr_range_int         ... bench:       3,107 ns/iter (+/- 12) = 2574 MB/s
```

After:
```
test distr_baseline          ... bench:       1,617 ns/iter (+/- 15) = 4947 MB/s
test distr_exp               ... bench:       6,090 ns/iter (+/- 121) = 1313 MB/s
test distr_gamma_large_shape ... bench:      17,419 ns/iter (+/- 215) = 459 MB/s
test distr_gamma_small_shape ... bench:      20,304 ns/iter (+/- 259) = 394 MB/s
test distr_log_normal        ... bench:       5,849 ns/iter (+/- 64) = 1367 MB/s
test distr_normal            ... bench:       5,854 ns/iter (+/- 65) = 1366 MB/s
test distr_range2_float      ... bench:       1,620 ns/iter (+/- 7) = 4938 MB/s
test distr_range2_int        ... bench:       3,098 ns/iter (+/- 12) = 2582 MB/s
test distr_range_float       ... bench:       1,618 ns/iter (+/- 8) = 4944 MB/s
test distr_range_int         ... bench:       3,100 ns/iter (+/- 11) = 2580 MB/s
```

The benchmarks that depend on the ziggurat method have all improved, `normal` and `log_normal` by 12%, `gamma` by 4.5%, and `exp` by 3%. The benchmark `distr_range2_float` does not regress, thanks to splitting the float conversion functions off from those also generating the random integers.

Thanks for the encouragement until now!