@pitdicker 2017-08-29
I just learned the `gamma` and derived distributions do not use the ziggurat method. That makes me wonder why they have improved. Maybe the new code for `Open01` is what makes them slightly faster...