@pitdicker 2017-08-31
The conversion functions are now moved to a different module. The new names make it clearer what is going on, so that is a plus :-)