@pitdicker 2017-09-03
Someday I will learn how to rebase properly...

I have made the changes you suggested.
I am done tweaking this code. Maybe in the future I will have another try at the rounding with ranges.