@dhardy 2017-09-03
Great, thanks. Something like `git rebase -i HEAD~5` is extremely useful for rebasing.